refactor to make pages server components
